requirements.txt
django==3.0.0
djangorestframework==3.11.0
django-environ==0.4.5
psycopg2-binary==2.8.3
django-cors-headers==3.2.1
django-smart-selects
gunicorn
model_mommy